% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptR_job.R
\name{print.summary.cryptR_job}
\alias{print.summary.cryptR_job}
\title{Compact Print Method for a \code{summary.cryptR_job}}
\usage{
\method{print}{summary.cryptR_job}(x, ...)
}
\arguments{
\item{x}{A \code{summary.cryptR_job} object.}

\item{...}{Ignored.}
}
\value{
\code{invisible(x)}.
}
\description{
Compact Print Method for a \code{summary.cryptR_job}
}
