% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt_data.R
\name{crypt_data}
\alias{crypt_data}
\title{Encrypt variables within a data frame}
\usage{
crypt_data(
  loaded_dataset,
  vars_to_encrypt,
  vars_to_remove = NULL,
  encryption_key,
  algorithm = "md5",
  correspondence_table = TRUE,
  correspondence_table_label = NULL
)
}
\arguments{
\item{loaded_dataset}{The dataset for which user wants
to encrypt at least one variable.
It must be an expression and not a character vector.}

\item{vars_to_encrypt}{Character vector of variables to encrypt.}

\item{vars_to_remove}{Character vector of variables to remove.}

\item{encryption_key}{Character vector.}

\item{algorithm}{Algorithm to use.
From digest::digest()'s algo argument.}

\item{correspondence_table}{Logical 1L. TRUE if user wants a correspondence table
between initial and encrypted values.}

\item{correspondence_table_label}{Character 1L. Label of the correspondence table.}
}
\description{
This function corresponds to the
in-session version of the encryption process.
For the considered loaded dataset, variables to encrypt can be
in any number, following users's needs.
}
\examples{
crypt_data(
  loaded_dataset = mtcars[1:5, ], 
  vars_to_encrypt = "mpg", 
  vars_to_remove = "cyl", 
  encryption_key = "1234567", 
  algorithm = "md5", 
  correspondence_table = FALSE
)
}
\seealso{
cryptR::crypt_r(), digest::digest().
}
