% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt_r.R
\name{crypt_r}
\alias{crypt_r}
\title{Industrialized dataset variables encryption.}
\usage{
crypt_r(
  mask_folder_path,
  mask_file,
  output_path,
  intermediate_path,
  encryption_key,
  algorithm = "md5",
  correspondence_table = TRUE
)
}
\arguments{
\item{mask_folder_path}{Folder path of the excel mask.}

\item{mask_file}{File name (with extension) of the excel mask.}

\item{output_path}{Folder path where the encrypted datasets
will be created.}

\item{intermediate_path}{Folder path where intermediate files
(correspondence tables for e.g) will be created.}

\item{encryption_key}{Character vector.}

\item{algorithm}{Algorithm to use.
From digest::digest()'s algo argument.}

\item{correspondence_table}{Logical vector. If TRUE, the correspondence tables will computed
and stored in the indicated intermediate_path.}
}
\description{
This function consolidates a set of procedures
made to encrypt variables from datasets.
Variables to encrypt and input datasets can be in any number,
following user's needs.
The design tries to approach parallel processing.
}
