% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypt_vector.R
\name{crypt_vector}
\alias{crypt_vector}
\title{Encrypt a vector}
\usage{
crypt_vector(vector, key, algo)
}
\arguments{
\item{vector}{Vector to process}

\item{key}{Character 1L. Encryption key}

\item{algo}{Character 1L. Algorithm.
See algo argument from digest::digest()}
}
\value{
Character. Encrypted vector
}
\description{
Encrypt a vector
}
\examples{
crypt_vector(vector = c("1234", "5678", "9101112", NA, ""), key = "123456", algo = "md5")
}
