% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptRopen-package.R
\docType{package}
\name{cryptRopen-package}
\alias{cryptRopen}
\alias{cryptRopen-package}
\title{cryptRopen: Industrialized Datasets Encryption (Open-Source Version)}
\description{
This package provides a framework to encrypt variables from datasets. Variables to encrypt and datasets can be in any number.
}
\author{
\strong{Maintainer}: Daniel Rakotomalala \email{rakdanielh@gmail.com}

}
\keyword{internal}
