% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect a data frame}
\usage{
inspect(data_frame, nrow = FALSE)
}
\arguments{
\item{data_frame}{The data.frame to explore.
Need to exist in the Global Environment.}

\item{nrow}{Logical. If TRUE, the number of observations of
the dataset is rendered in addition.}
}
\value{
Variable list of the dataset and
systematic informations for each variable.
}
\description{
Given a data frame as an input, this function outputs an broad view
of this data frame contents:
dimensions, list of variables and
systematic corresponding information (type,
frequence and proportion of unique values, missing and 0-length values;
and up to the 10 first unique values).
}
\examples{
requireNamespace("magrittr")
inspect(CO2)
}
